requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
deprecation==2.1.0
ecdsa==0.18.0
idna==3.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.3.0
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==1.26.6
websocket-client==1.6.1
xarray==2023.6.0
